Make perf scripts run without installing megatron-bridge (#760)
* draft #1


* temp fixes


* revert temp code


* revert


* minor fix


* lint


* lint


* fix lint


* adding copyright


---------
