+++ scripts/performance/utils/common.py
@@ -1,0 +1,30 @@
+# Copyright (c) 2025, NVIDIA CORPORATION.  All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from typing import Any
+
+
+def get_perf_matrix_overrides(yaml_root: Any, args: Any) -> Any:
+    """Get the performance matrix overrides from the YAML file."""
+    perf = yaml_root.get("perf_matrix") if hasattr(yaml_root, "get") else None
+    if not perf:
+        return
+    if args.gpu not in perf:
+        return
+    num_gpus_value = args.num_gpus or args.gpus_per_node
+    num_gpus_yaml_key = f"num_gpus_{num_gpus_value}"
+    gpu_block = perf.get(args.gpu) or {}
+    preset = gpu_block.get(num_gpus_yaml_key) or {}
+
+    return preset
